+++ .github/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Improvements
 
+* Improves the efficiency of Hermite polynomial calculation in `hermite_multidimensional.hpp`. [#141](https://github.com/XanaduAI/thewalrus/pull/141)
+
 ### Bug fixes
 
 ### Contributors