+++ .github/CHANGELOG.md
@@ -11,6 +11,8 @@
 
 ### Bug fixes
 
+* Corrects the issue with hbar taking a default value when calling `state_vector` and `density_matrix_element` [#149] (https://github.com/XanaduAI/thewalrus/pull/149)
+
 ### Contributors
 
 This release contains contributions from (in alphabetical order):